d_a_obj_figure OK for demo, fix wrong enum usage for retail

+++ include/f_pc/f_pc_manager.h
@@ -80,7 +80,7 @@ inline base_process_class* fpcM_Search(fpcLyIt_JudgeFunc i_func, void* i_data) {
     return fpcEx_Search(i_func, i_data);
 }
 
-inline s8 fpcM_CreateResult(void* pProc) {
+inline int fpcM_CreateResult(void* pProc) {
     return ((base_process_class*)pProc)->mCreateResult;
 }
